+++ formio/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+Implement "Form Ref" field on several models regarding assets and attachments:\
+`formio.version.asset`, `formio.extra.asset`, `ir.attachment`\
+This facilitates purposes like export/import tools.
+
 ## 12.4
 
 Improve form builder wizard save as draft: previous page, page clicked.