+++ formio/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 16.0.1.4
 
-Implement "Form Ref" field on several models regarding assets and attachments:\
+Implement "Forms Ref" field on several models regarding assets and attachments:\
 `formio.version.asset`, `formio.extra.asset`, `ir.attachment`\
 This facilitates purposes like export/import tools.